Enforce deployment pipeline
Create two new commands to enforce the expected deployment pipeline.

`deploy-latest-to-staging` will deploy the most recent version label
that starts with `release-`. Ultimately these will only be created on
successfull completion of a CI run on master.

`deploy-staging-to-production` deploys the version label that is
currently deployed to staging into production.

The two commands that allow any version label to be deployed into
staging and production have now been marked as dangerous. The idea is
that work should flow linearly through CI and staging before ending up
in production.