Add an RDS instance per Beanstalk environment
This is a very large commit, sorry. The full details of the process has
been added to the README so I will just outline what has changed here.

When a new Beanstalk environment is created an RDS instace is also
created. The Beanstalk environment is given acess to this with a
VPC security group. A Beanstalk configuration template is also created
that contains the RDS connection information (in the form of an
SQLAlchemy URI). This is used when creating the Beanstalk environment.

The order of creation is very significant, it goes like this:

- Create RDS security group
- Create RDS instance
- Wait for RDS instance to come up
- Create configuration template with RDS connection details
- Create Beanstalk environment
- Wait for Beanstal environment security group
- Allow access to RDS instance security group from Beanstalk
  environment security group.

We need to wait for the RDS instance to come up so that we can get the
endpoint it lives on to give to the Beanstalk environment. We need to
wait for Beanstalk security group in order to give it access to the RDS
instance. Note that there is a race condition; if the Beanstalk
environment runs the application startup scripts that do database
migration before we've updated the RDS security group it will fail.
Luckily the Beanstalk environment security group seems to be created
very early in the process.

The RDS instance and security group both have 'db-' prefixed to them.
This is because they must start with a letter and the app sha may start
with a number.