Set 'RDS_' environment variables
When Beanstalk attaches an RDS instance to an environment automatically
it makes certain environment variables available. I'm making them
available here to be as close to how Beanstalk works as possible. The
reason for this is to make it easier to modify this tool so that it's
not Digital Marketplace specific.

See:
http://docs.aws.amazon.com/elasticbeanstalk/latest/dg/create_deploy_Python.rds.html#create_deploy_Python.rds.newDB