Rename ephemeral environments to development
It wasn't clear what these environments were actually for. Also, because
these environments are a bit more costly to create now that we have an
RDS instance attached to each environment it probably doesn't make sense
to have the binding to feature branches quite so tight.

Now a development environment must be given a name when it's created.
Nothing else about it being on a branch is assumed.